# api/api_specification.yaml
openapi: 3.0.3
info:
  title: API Specification
  description: API specification for the Instagram Clone API
  version: 0.0.1
  contact:
    name: NovqiGarrix
    url: https://twitter.com/NovqiGarrix
servers:
  - url: 'http://localhost:4000/api/v1'
    description: Localhost server
  - url: 'https://igclone.fly.dev/api/v1'
    description: Fly.io server
security:
  - jwt: [ ]
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    BadRequestError:
      type: object
      properties:
        code:
          type: integer
          default: 400
        errors:
          type: array
          items:
            type: object
            properties:
              error:
                type: string
                nullable: true
                description: Error message if there is any
              message:
                type: string
                nullable: true
                description: Error message due to validation
              field:
                type: string
                nullable: true
                description: Field that cause the error
    UnauthorizedError:
      type: object
      properties:
        code:
          type: integer
          default: 401
        errors:
          type: array
          items:
            type: object
            properties:
              error:
                type: string
...
    InternalServerError:
      type: object
      properties:
        code:
          type: integer
          default: 500
        errors:
          type: object
          properties:
            error:
              type: string
...
    SimpleUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        username:
          type: string
        pictureUrl:
          type: string
...
  responses:
    401:
      description: Unauthorized. Could be because of missing or invalid or expired jwt token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedError'
...
    500:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerError'
paths:
  "/auth/signup":
    post:
      tags:
        - Auth API
      summary: This endpoint is used to register a new user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                fullName:
                  type: string
                username:
                  type: string
                password:
                  type: string
        required: true
      responses:
        201:
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    default: 201

        400:
          description: Bad Request. Could be because of missing required request body or invalid data on request body or the email already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'

        500:
          $ref: '#/components/responses/500'
  "/auth":
    post:
      tags:
        - Auth API
      summary: This endpoint is used to log in a user and get the jwt token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                emailUsername:
                  type: string
                password:
                  type: string
        required: true
      responses:
        200:
          description: User logged in successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    default: 200
                  data:
                    $ref: '#/components/schemas/SimpleUser'
                  token:
                    type: string
                  refreshToken:
                    type: string

        400:
          description: Bad Request. Could be because of missing required request body or invalid request body or invalid email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'

        500:
          $ref: '#/components/responses/500'
  "/auth/me":
    get:
      tags:
        - Auth API
      security:
        - jwt: [ ]
      summary: This endpoint is used to get the current logged in user
      responses:
        200:
          description: Successfully retrieved the current logged in user
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    default: 200
                  data:
                    $ref: '#/components/schemas/SimpleUser'

        401:
          $ref: '#/components/responses/401'

        500:
          $ref: '#/components/responses/500'
  "/auth/token":
    post:
      description: This endpoint is used to get a new jwt token using the refresh token
      tags:
        - Auth API
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        200:
          description: Successfully retrieved a new jwt token
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    default: 200
                  token:
                    type: string

        401:
          description: Your refresh token is invalid or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'

        500:
          $ref: '#/components/responses/500'
...